requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
pytest==9.0.1
httpx==0.28.1
requests==2.32.5
python-dotenv==1.2.1
pyjwt==2.10.1
asyncpg==0.31.0
pytest-asyncio==1.3.0
